+++ src/modules/mo/mo_dhl/Application/views/admin/de/module_options.php
@@ -42,10 +42,10 @@ $aLang = [
     'SHOP_MODULE_mo_dhl__notification_mode_ALWAYS'             => 'Immer',
     'SHOP_MODULE_mo_dhl__notification_mode_ASK'                => 'Auf Kundenwunsch übermitteln',
 
-    'SHOP_MODULE_mo_dhl__ident_check_min_age'     => 'Mindestalter für den IdentCheck',
-    'SHOP_MODULE_mo_dhl__ident_check_min_age_0'   => 'kein Mindestalter',
-    'SHOP_MODULE_mo_dhl__ident_check_min_age_A16' => '16 Jahre',
-    'SHOP_MODULE_mo_dhl__ident_check_min_age_A18' => '18 Jahre',
+    'SHOP_MODULE_mo_dhl__ident_check_min_age'    => 'Mindestalter für den IdentCheck',
+    'SHOP_MODULE_mo_dhl__ident_check_min_age_0'  => 'kein Mindestalter',
+    'SHOP_MODULE_mo_dhl__ident_check_min_age_16' => '16 Jahre',
+    'SHOP_MODULE_mo_dhl__ident_check_min_age_18' => '18 Jahre',
 
     'SHOP_MODULE_GROUP_mo_dhl__retoure'                            => 'Retoure',
     'SHOP_MODULE_mo_dhl__retoure_reference_prefix'                 => 'Präfix für die Sendungsreferenz',
@@ -57,6 +57,8 @@ $aLang = [
     'SHOP_MODULE_mo_dhl__retoure_allow_frontend_creation_ALWAYS'   => 'Immer',
     'SHOP_MODULE_mo_dhl__retoure_days_limit'                       => 'Max. Zeitraum zur Erstellung von Retouren (in Tagen)',
     'HELP_SHOP_MODULE_mo_dhl__retoure_days_limit'                  => 'Kunden können nur innerhalb einer bestimmten Zeit ab Versand der Ware ein Retourenlabel erstellen.',
+    'SHOP_MODULE_mo_dhl__retoure_admin_approve'                    => 'Kunden sollten nach einem Retourenlabel fragen',
+    'HELP_SHOP_MODULE_mo_dhl__retoure_admin_approve'               => 'Sie können angeben, dass Kunden Retourenlabels nicht selbst erstellen dürfen, sondern nur die Erstellung anfragen können. Diese muss dann im Adminbereich selbst erstellt werden.',
 
     'SHOP_MODULE_GROUP_mo_dhl__beilegerretoure'            => 'Beilegerretoure',
     'SHOP_MODULE_mo_dhl__beilegerretoure_active'           => 'Beilegerretoure aktiv',