+++ src/modules/mo/mo_dhl/Application/views/admin/en/module_lang.php
@@ -24,6 +24,10 @@ $aLang = [
     'MO_DHL__PARTICIPATION_NUMBER_ERROR'           => 'A participation number is exactly two characters long and consists solely of letters and digits',
     'MO_DHL__OPERATOR'                             => 'Delivery operator',
     'HELP_MO_DHL__OPERATOR'                        => 'Please provide the name of the delivery operator. This information is needed for a possible return outside the EU.',
+    'MO_DHL__CUSTOMER_RETOURE_REQUEST_STATUS'      => 'Customer retoure request status',
+    'MO_DHL__REQUESTED'                            => 'Requested',
+    'MO_DHL__CREATED'                              => 'Created',
+    'MO_DHL__DECLINED'                             => 'Declined',
     'MO_DHL__PROCESS_IDENTIFIER'                   => 'Process number',
     'MO_DHL__PROCESS_IDENTIFIER_ERROR'             => 'Please choose a process number from the list.',
     'MO_DHL__DELIVERYSET_DHL'                      => 'DHL',
@@ -44,6 +48,7 @@ $aLang = [
     'MO_DHL__INSTALL_FOLDER_ERROR'                 => 'The directory %s could not be created. Please create it manually',
     'MO_DHL__CREATE_LABEL'                         => 'Create a new delivery label',
     'MO_DHL__CREATE_LABELS'                        => 'Create labels',
+    'MO_DHL__CREATE_RETOURE_LABELS'                => 'Create Retoure labels',
     'MO_DHL__DELETE_SHIPMENT'                      => 'Cancel shipment',
 
     'MO_DHL__CUSTOM_LABEL_CREATE'             => 'Create individual label',
@@ -74,12 +79,13 @@ $aLang = [
 
     'MO_DHL__ARTICLES_DHL'                => 'DHL',
     'MO_DHL__CATEGORIES_DHL'              => 'DHL',
+    'MO_DHL__VISUAL_AGE_CHECK'            => 'Age check',
     'MO_DHL__VISUAL_AGE_CHECK16'          => 'Age check: 16 years',
     'MO_DHL__VISUAL_AGE_CHECK18'          => 'Age check: 18 years',
     'MO_DHL__BULKY_GOOD'                  => 'Bulky good',
     'MO_DHL__IDENT_CHECK'                 => 'Ident-Check',
+    'MO_DHL__FOR_IDENT_CHECK'             => 'for Ident-Check',
     'MO_DHL__CASH_ON_DELIVERY'            => 'Cash on delivery',
     'MO_DHL__ADDITIONAL_INSURANCE'        => 'Additional Insurance',
     'HELP_MO_DHL__ADDITIONAL_INSURANCE'   => 'DHL normally insures the delivery inside Germany up to 500 EUR concerning losing or damaging. For deliveries of more expensive products DHL offers an additional insurance up to 2.500 EUR (+6,00 EUR surcharge) or 25.000 EUR  (+18,00 EUR surcharge). You can find more informationens <a href="https://www.dhl.de/content/dam/images/pdf/GK/Services/dhl-transportversicherung-infoblatt-en-052020.pdf">here</a>',
-
 ];