+++ src/modules/mo/mo_dhl/Install.php
@@ -179,6 +179,7 @@ class Install
         $order = self::addColumn('oxorder', 'MO_DHL_EKP', 'CHAR(10)')
             + self::addColumn('oxorder', 'MO_DHL_PROCESS', 'VARCHAR(32)')
             + self::addColumn('oxorder', 'MO_DHL_OPERATOR', 'VARCHAR(40)')
+            + self::addColumn('oxorder', 'MO_DHL_RETOURE_REQUEST_STATUS', 'VARCHAR(32)')
             + self::addColumn('oxorder', 'MO_DHL_PARTICIPATION', 'CHAR(2)')
             + self::addColumn('oxorder', 'MO_DHL_LAST_LABEL_CREATION_STATUS', 'VARCHAR(100)')
             + self::addColumn('oxorder', 'MO_DHL_ALLOW_NOTIFICATION', 'TINYINT(1) NOT NULL DEFAULT 0')